+++ test/flow.test.ts
@@ -699,6 +699,10 @@ describe('event flow use cases', () => {
     expect(map.getElement('original.boolean_value')).toBe(true);
     expect(map.getElement('original.negate_value')).toBe(false);
     expect(map.getElement('original.double_negate_value')).toBe(true);
+    expect(map.getElement('original.none_is_true')).toBe(true);
+    expect(map.getElement('original.none_is_false')).toBe(false);
+    expect(map.getElement('original.unique_id1')).not.toEqual(map.getElement('original.unique_id2'));
+    expect(map.getElement('original.unique_id2')).toEqual(map.getElement('original.unique_id3'));
     expect(map.getElement('demo1')).toBe("ok");
     expect(map.getElement('demo2')).toBe(testUser);
     expect(map.getElement('demo3')).toBe("greetings");